+++ tests/API/Function/Conversation/ShowTest.php
@@ -7,40 +7,22 @@ use Friendica\Test\API\ApiTest;
 class ShowTest extends ApiTest
 {
 	/**
-	 * Test the api_conversation_show() function.
+	 * Test the api_blocks_list() function.
 	 * @return void
-	 * @expectedException Friendica\Network\HTTPException\BadRequestException
 	 */
 	public function testDefault()
 	{
-		api_conversation_show('json');
+		$result = api_blocks_list('json');
+		$this->assertArrayHasKey('user', $result);
 	}
 
 	/**
-	 * Test the api_conversation_show() function with an ID.
+	 * Test the api_blocks_list() function an undefined cursor GET variable.
 	 * @return void
 	 */
-	public function testWithId()
+	public function testWithUndefinedCursor()
 	{
-		$this->app->argv[3] = 1;
-		$_REQUEST['max_id'] = 10;
-		$_REQUEST['page'] = -2;
-		$result = api_conversation_show('json');
-		$this->assertNotEmpty($result['status']);
-		foreach ($result['status'] as $status) {
-			$this->assertStatus($status);
-		}
-	}
-
-	/**
-	 * Test the api_conversation_show() function with an unallowed user.
-	 * @return void
-	 * @expectedException Friendica\Network\HTTPException\ForbiddenException
-	 */
-	public function testWithUnallowedUser()
-	{
-		$_SESSION['allow_api'] = false;
-		$_GET['screen_name'] = $this->selfUser['nick'];
-		api_conversation_show('json');
+		$_GET['cursor'] = 'undefined';
+		$this->assertFalse(api_blocks_list('json'));
 	}
 }
