+++ tests/API/Function/OAuth/AccessTokenTest.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Friendica\Test\API\OAuth;
+
+use Friendica\Test\API\ApiTest;
+
+class AccessTokenTest extends ApiTest
+{
+	/**
+	 * Test the api_oauth_access_token() function.
+	 * @return void
+	 */
+	public function testDefault()
+	{
+		$this->markTestIncomplete('killme() kills phpunit as well');
+	}
+}
