+++ tests/API/Function/OAuth/RequestTokenTest.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Friendica\Test\API\OAuth;
+
+use Friendica\Test\API\ApiTest;
+
+class RequestTokenTest extends ApiTest
+{
+	/**
+	 * Test the api_oauth_request_token() function.
+	 * @return void
+	 */
+	public function testDefault()
+	{
+		$this->markTestIncomplete('killme() kills phpunit as well');
+	}
+}
