+++ tests/API/Util/Format/FormatDataTest.php
