+++ tests/API/Util/Format/FormatItemsActivitiesTest.php
