+++ tests/API/Util/Format/FormatItemsEbededImagesTest.php
