+++ tests/API/Util/Format/FormatItemsTest.php
