+++ tests/API/Util/Format/FormatMessagesTest.php
