+++ tests/API/Util/PostPhotoItemTest.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Friendica\Test\API;
+
+class PostPhotoItemTest extends ApiTest
+{
+	/**
+	 * Test the post_photo_item() function.
+	 * @return void
+	 */
+	public function testDefault()
+	{
+		$this->markTestIncomplete();
+	}
+}
