+++ tests/API/Util/PreparePhotoDataTest.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Friendica\Test\API;
+
+class PreparePhotoDataTest extends ApiTest
+{
+	/**
+	 * Test the prepare_photo_data() function.
+	 * @return void
+	 */
+	public function testDefault()
+	{
+		$this->markTestIncomplete();
+	}
+}
