+++ tests/API/Util/SaveMediaToDatabaseTest.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Friendica\Test\API;
+
+class SaveMediaToDatabaseTest extends ApiTest
+{
+	/**
+	 * Test the save_media_to_database() function.
+	 * @return void
+	 */
+	public function testDefault()
+	{
+		$this->markTestIncomplete();
+	}
+}
